Create chessboard_check.py
This program tries to evaluate the validity of a given chessboard using the following criteria:
1. There must be exactly one black and one white king on the board. 
2. Each player must have at most 16 pieces. 
3. Each player must have at most 8 pawns. 
4. All the pieces must be on a valid square from '1a' to '8h'. 
5. The pieces must begin with either a 'w' for white or 'b' for black pieces and must be followed by a piece name (pawn, bishop, knight, rook, queen, king).


NOTE: The board to be checked is to presented in the form a dictionary. For example, if we have the white king on '1a', black king on '3d' and a black pawn on '4c'. The cheesbaord should be represented as shown below. 

sampleBoard = {'wking': '1a', 'bking': '3d', 'bpawn': '4c'}.