Create coin_flip_streaks.py
A program that tries to determine the the probability that 6 consecutive heads or tails will occur in a series of 100 coin flips. The probability might seem really low to humans, but in reality, the occurrence is highly probable as will be demonstrated in the program. 

Try out it (you'll be surprised, or maybe not if you're good at statistics).